refactor(mobile/flutter): service details and services pages with new widgets for better structure and readability
