+++ mobile/lib/pages/service_details_page.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
 import '../models/service.dart';
+import '../widgets/service_header_card.dart';
+import '../widgets/action_reaction_widgets.dart';
+import '../widgets/state_widgets.dart';
 
 class ServiceDetailsPage extends StatelessWidget {
   final Service service;
@@ -22,209 +25,39 @@ class ServiceDetailsPage extends StatelessWidget {
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             // Service Header
-            Card(
-              elevation: 4,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(12),
-              ),
-              child: Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: Row(
-                  children: [
-                    Container(
-                      width: 64,
-                      height: 64,
-                      decoration: BoxDecoration(
-                        color: Theme.of(context).primaryColor.withValues(alpha: 0.1),
-                        borderRadius: BorderRadius.circular(12),
-                      ),
-                      child: Icon(
-                        _getServiceIcon(service.name),
-                        color: Theme.of(context).primaryColor,
-                        size: 32,
-                      ),
-                    ),
-                    const SizedBox(width: 16),
-                    Expanded(
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Text(
-                            service.displayName,
-                            style: Theme.of(context).textTheme.headlineSmall?.copyWith(
-                              fontWeight: FontWeight.bold,
-                            ),
-                          ),
-                          const SizedBox(height: 4),
-                          Text(
-                            '${service.actions.length} actions • ${service.reactions.length} reactions',
-                            style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                              color: Colors.grey[600],
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-
+            ServiceHeaderCard(service: service),
             const SizedBox(height: 24),
 
             // Actions Section
-            if (service.actions.isNotEmpty) ...[
-              Text(
-                'Actions',
-                style: Theme.of(context).textTheme.titleLarge?.copyWith(
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              const SizedBox(height: 12),
-              ...service.actions.map((action) => _buildActionReactionCard(
-                context,
-                action.displayName,
-                action.description,
-                Icons.play_arrow,
-                Colors.green,
-              )),
-              const SizedBox(height: 24),
-            ],
+            ActionReactionSection(
+              title: 'Actions',
+              items: service.actions.map((action) => ActionReactionCard(
+                title: action.displayName,
+                description: action.description,
+                icon: Icons.play_arrow,
+                iconColor: Colors.green,
+              )).toList(),
+            ),
 
             // Reactions Section
-            if (service.reactions.isNotEmpty) ...[
-              Text(
-                'Reactions',
-                style: Theme.of(context).textTheme.titleLarge?.copyWith(
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              const SizedBox(height: 12),
-              ...service.reactions.map((reaction) => _buildActionReactionCard(
-                context,
-                reaction.displayName,
-                reaction.description,
-                Icons.flash_on,
-                Colors.orange,
-              )),
-            ],
+            ActionReactionSection(
+              title: 'Reactions',
+              items: service.reactions.map((reaction) => ActionReactionCard(
+                title: reaction.displayName,
+                description: reaction.description,
+                icon: Icons.flash_on,
+                iconColor: Colors.orange,
+              )).toList(),
+            ),
 
             // Empty State
             if (service.actions.isEmpty && service.reactions.isEmpty)
-              Center(
-                child: Padding(
-                  padding: const EdgeInsets.symmetric(vertical: 48.0),
-                  child: Column(
-                    children: [
-                      Icon(
-                        Icons.info_outline,
-                        size: 64,
-                        color: Colors.grey[400],
-                      ),
-                      const SizedBox(height: 16),
-                      Text(
-                        'No actions or reactions available',
-                        style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-                          color: Colors.grey[600],
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
+              const EmptyStateWidget(
+                message: 'No actions or reactions available',
               ),
           ],
         ),
       ),
     );
-  }
-
-  Widget _buildActionReactionCard(
-    BuildContext context,
-    String title,
-    String description,
-    IconData icon,
-    Color iconColor,
-  ) {
-    return Card(
-      elevation: 2,
-      margin: const EdgeInsets.only(bottom: 8.0),
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(8),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.all(12.0),
-        child: Row(
-          children: [
-            Container(
-              width: 40,
-              height: 40,
-              decoration: BoxDecoration(
-                color: iconColor.withValues(alpha: 0.1),
-                borderRadius: BorderRadius.circular(8),
-              ),
-              child: Icon(
-                icon,
-                color: iconColor,
-                size: 20,
-              ),
-            ),
-            const SizedBox(width: 12),
-            Expanded(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    title,
-                    style: Theme.of(context).textTheme.titleSmall?.copyWith(
-                      fontWeight: FontWeight.w600,
-                    ),
-                  ),
-                  const SizedBox(height: 2),
-                  Text(
-                    description,
-                    style: Theme.of(context).textTheme.bodySmall?.copyWith(
-                      color: Colors.grey[600],
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  IconData _getServiceIcon(String serviceName) {
-    switch (serviceName.toLowerCase()) {
-      case 'facebook':
-        return Icons.facebook;
-      case 'twitter':
-      case 'x':
-        return Icons.alternate_email;
-      case 'gmail':
-      case 'email':
-        return Icons.email;
-      case 'discord':
-        return Icons.chat;
-      case 'github':
-        return Icons.code;
-      case 'spotify':
-        return Icons.music_note;
-      case 'youtube':
-        return Icons.play_circle_fill;
-      case 'slack':
-        return Icons.message;
-      case 'telegram':
-        return Icons.send;
-      case 'weather':
-        return Icons.wb_sunny;
-      case 'timer':
-        return Icons.timer;
-      case 'rss':
-        return Icons.rss_feed;
-      default:
-        return Icons.apps;
-    }
   }
 }
